add wasm propagate on main thread
