--- wasm/sgp4-wrap/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "sgp4_wrap"
-version = "0.0.0"
-edition = "2021"
-
-[lib]
-crate_type = ["cdylib"]
-
-[dependencies]
-sgp4 = "0.7"
-eventual = "0.1.7"
-serde = { version = "1.0", features = ["derive"] }
-wasm-bindgen = { version = "=0.2.81", features = ["serde-serialize"] }
